+++ src/pages/content/_id/index.vue
@@ -12,10 +12,10 @@
 </template>
 
 <script>
+import ContentModule from "@/store/content";
 import LernstoreDetailView from "@components/organisms/LernstoreDetailView";
 import LernstoreCollectionDetailView from "@components/organisms/LernstoreCollectionDetailView";
 import BaseSpinner from "@components/base/BaseSpinner";
-import { mapGetters } from "vuex";
 
 export default {
 	meta: {
@@ -26,25 +26,28 @@ export default {
 		LernstoreCollectionDetailView,
 		BaseSpinner,
 	},
-	layout({ store, query }) {
+	layout({ query }) {
 		return String(query.isCollection) === "true" &&
-			store.getters["content/getCollectionsFeatureFlag"] === true
+			ContentModule.getCollectionsFeatureFlag === true
 			? "loggedInFull"
 			: "plain";
 	},
 	computed: {
-		...mapGetters("content", {
-			resource: "getCurrentResource",
-			collectionsFeatureFlag: "getCollectionsFeatureFlag",
-			status: "getStatus",
-			isCollection: "isCollection",
-		}),
+		resource() {
+			return ContentModule.getCurrentResource;
+		},
+		collectionsFeatureFlag() {
+			return ContentModule.getCollectionsFeatureFlag;
+		},
+		status() {
+			return ContentModule.getStatus;
+		},
+		isCollection() {
+			return ContentModule.isCollection;
+		},
 	},
 	async created() {
-		await this.$store.dispatch(
-			"content/getResourceMetadata",
-			this.$route.params.id
-		);
+		await ContentModule.getResourceMetadata(this.$route.params.id);
 	},
 };
 </script>