+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Changed
 
+- BC-5 - renamed sidebar item "Open Tasks" to "Current Tasks"
+- BC-18 - refactored the envConfig, accounts, autoLogout, and filePaths store modules to typescript and the corresponding components' access to them
 - BC-14 - changed thr contact address
 - BC-64 - execute e2e tests via push on main
 - BC-37 - BC-56 - reduce resource consumption for deployed nuxtclient
@@ -22,6 +24,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 - SC-9192 - access feathers and nestjs apps over their dedicated version prefixes
 - SC-9191 run e2e tests in pull requests with 'run tests' label
+- SC-9280 fix primary and secondary colors according to old client
 
 ## [26.8.0] - 2021-08-05